@@ -3,8 +3,6 @@ using Neo.IO.Json;
 using Neo.Ledger;
 using Neo.Persistence;
 using Neo.SmartContract.Native;
-using Neo.SmartContract.Native.Oracle;
-using Neo.SmartContract.Native.Designate;
 using Neo.VM;
 using System;
 using System.IO;
@@ -33,7 +31,7 @@ namespace Neo.Network.P2P.Payloads
         static OracleResponse()
         {
             using ScriptBuilder sb = new ScriptBuilder();
-            sb.EmitAppCall(NativeContract.Oracle.Hash, "finish");
+            sb.EmitDynamicCall(NativeContract.Oracle.Hash, "finish", false);
             FixedScript = sb.ToArray();
         }
 
@@ -71,7 +69,7 @@ namespace Neo.Network.P2P.Payloads
             OracleRequest request = NativeContract.Oracle.GetRequest(snapshot, Id);
             if (request is null) return false;
             if (tx.NetworkFee + tx.SystemFee != request.GasForResponse) return false;
-            UInt160 oracleAccount = Blockchain.GetConsensusAddress(NativeContract.Designate.GetDesignatedByRole(snapshot, Role.Oracle, snapshot.Height + 1));
+            UInt160 oracleAccount = Blockchain.GetConsensusAddress(NativeContract.RoleManagement.GetDesignatedByRole(snapshot, Role.Oracle, snapshot.Height + 1));
             return tx.Signers.Any(p => p.Account.Equals(oracleAccount));
         }
     }
