@@ -4,12 +4,13 @@ namespace Neo.Network.P2P.Payloads
     {
         Success = 0x00,
 
-        ConsensusUnreachable = 0x10,
-        NotFound = 0x12,
-        Timeout = 0x14,
-        Forbidden = 0x16,
-        ResponseTooLarge = 0x18,
-        InsufficientFunds = 0x1a,
+        ProtocolNotSupported = 0x10,
+        ConsensusUnreachable = 0x12,
+        NotFound = 0x14,
+        Timeout = 0x16,
+        Forbidden = 0x18,
+        ResponseTooLarge = 0x1a,
+        InsufficientFunds = 0x1c,
 
         Error = 0xff
     }
