@@ -1,5 +1,5 @@
-using System.Collections.Generic;
 using Neo.Network.P2P.Payloads;
+using System.Collections.Generic;
 
 namespace Neo.Plugins
 {
