@@ -1,4 +1,3 @@
-using Neo.SmartContract.Enumerators;
 using Neo.SmartContract.Iterators;
 using Neo.VM.Types;
 using System;
@@ -8,42 +7,30 @@ namespace Neo.SmartContract
 {
     partial class ApplicationEngine
     {
-        public static readonly InteropDescriptor System_Iterator_Create = Register("System.Iterator.Create", nameof(CreateIterator), 0_00000400, CallFlags.None, false);
-        public static readonly InteropDescriptor System_Iterator_Key = Register("System.Iterator.Key", nameof(IteratorKey), 0_00000400, CallFlags.None, false);
-        public static readonly InteropDescriptor System_Iterator_Keys = Register("System.Iterator.Keys", nameof(IteratorKeys), 0_00000400, CallFlags.None, false);
-        public static readonly InteropDescriptor System_Iterator_Values = Register("System.Iterator.Values", nameof(IteratorValues), 0_00000400, CallFlags.None, false);
-        public static readonly InteropDescriptor System_Iterator_Concat = Register("System.Iterator.Concat", nameof(ConcatIterators), 0_00000400, CallFlags.None, false);
+        public static readonly InteropDescriptor System_Iterator_Create = Register("System.Iterator.Create", nameof(CreateIterator), 1 << 4, CallFlags.None);
+        public static readonly InteropDescriptor System_Iterator_Next = Register("System.Iterator.Next", nameof(IteratorNext), 1 << 15, CallFlags.None);
+        public static readonly InteropDescriptor System_Iterator_Value = Register("System.Iterator.Value", nameof(IteratorValue), 1 << 4, CallFlags.None);
 
         protected internal IIterator CreateIterator(StackItem item)
         {
             return item switch
             {
                 Array array => new ArrayWrapper(array),
-                Map map => new MapWrapper(map),
+                Map map => new MapWrapper(map, ReferenceCounter),
                 VM.Types.Buffer buffer => new ByteArrayWrapper(buffer),
                 PrimitiveType primitive => new ByteArrayWrapper(primitive),
                 _ => throw new ArgumentException()
             };
         }
 
-        protected internal PrimitiveType IteratorKey(IIterator iterator)
+        protected internal bool IteratorNext(IIterator iterator)
         {
-            return iterator.Key();
+            return iterator.Next();
         }
 
-        protected internal IEnumerator IteratorKeys(IIterator iterator)
+        protected internal StackItem IteratorValue(IIterator iterator)
         {
-            return new IteratorKeysWrapper(iterator);
-        }
-
-        protected internal IEnumerator IteratorValues(IIterator iterator)
-        {
-            return new IteratorValuesWrapper(iterator);
-        }
-
-        protected internal IIterator ConcatIterators(IIterator first, IIterator second)
-        {
-            return new ConcatenatedIterator(first, second);
+            return iterator.Value();
         }
     }
 }
