@@ -7,190 +7,191 @@ namespace Neo.SmartContract
     {
         public static readonly IReadOnlyDictionary<OpCode, long> OpCodePrices = new Dictionary<OpCode, long>
         {
-            [OpCode.PUSHINT8] = 30,
-            [OpCode.PUSHINT16] = 30,
-            [OpCode.PUSHINT32] = 30,
-            [OpCode.PUSHINT64] = 30,
-            [OpCode.PUSHINT128] = 120,
-            [OpCode.PUSHINT256] = 120,
-            [OpCode.PUSHA] = 120,
-            [OpCode.PUSHNULL] = 30,
-            [OpCode.PUSHDATA1] = 180,
-            [OpCode.PUSHDATA2] = 13000,
-            [OpCode.PUSHDATA4] = 110000,
-            [OpCode.PUSHM1] = 30,
-            [OpCode.PUSH0] = 30,
-            [OpCode.PUSH1] = 30,
-            [OpCode.PUSH2] = 30,
-            [OpCode.PUSH3] = 30,
-            [OpCode.PUSH4] = 30,
-            [OpCode.PUSH5] = 30,
-            [OpCode.PUSH6] = 30,
-            [OpCode.PUSH7] = 30,
-            [OpCode.PUSH8] = 30,
-            [OpCode.PUSH9] = 30,
-            [OpCode.PUSH10] = 30,
-            [OpCode.PUSH11] = 30,
-            [OpCode.PUSH12] = 30,
-            [OpCode.PUSH13] = 30,
-            [OpCode.PUSH14] = 30,
-            [OpCode.PUSH15] = 30,
-            [OpCode.PUSH16] = 30,
-            [OpCode.NOP] = 30,
-            [OpCode.JMP] = 70,
-            [OpCode.JMP_L] = 70,
-            [OpCode.JMPIF] = 70,
-            [OpCode.JMPIF_L] = 70,
-            [OpCode.JMPIFNOT] = 70,
-            [OpCode.JMPIFNOT_L] = 70,
-            [OpCode.JMPEQ] = 70,
-            [OpCode.JMPEQ_L] = 70,
-            [OpCode.JMPNE] = 70,
-            [OpCode.JMPNE_L] = 70,
-            [OpCode.JMPGT] = 70,
-            [OpCode.JMPGT_L] = 70,
-            [OpCode.JMPGE] = 70,
-            [OpCode.JMPGE_L] = 70,
-            [OpCode.JMPLT] = 70,
-            [OpCode.JMPLT_L] = 70,
-            [OpCode.JMPLE] = 70,
-            [OpCode.JMPLE_L] = 70,
-            [OpCode.CALL] = 22000,
-            [OpCode.CALL_L] = 22000,
-            [OpCode.CALLA] = 22000,
-            [OpCode.ABORT] = 30,
-            [OpCode.ASSERT] = 30,
-            [OpCode.THROW] = 22000,
-            [OpCode.TRY] = 100,
-            [OpCode.TRY_L] = 100,
-            [OpCode.ENDTRY] = 100,
-            [OpCode.ENDTRY_L] = 100,
-            [OpCode.ENDFINALLY] = 100,
+            [OpCode.PUSHINT8] = 1 << 0,
+            [OpCode.PUSHINT16] = 1 << 0,
+            [OpCode.PUSHINT32] = 1 << 0,
+            [OpCode.PUSHINT64] = 1 << 0,
+            [OpCode.PUSHINT128] = 1 << 2,
+            [OpCode.PUSHINT256] = 1 << 2,
+            [OpCode.PUSHA] = 1 << 2,
+            [OpCode.PUSHNULL] = 1 << 0,
+            [OpCode.PUSHDATA1] = 1 << 3,
+            [OpCode.PUSHDATA2] = 1 << 9,
+            [OpCode.PUSHDATA4] = 1 << 12,
+            [OpCode.PUSHM1] = 1 << 0,
+            [OpCode.PUSH0] = 1 << 0,
+            [OpCode.PUSH1] = 1 << 0,
+            [OpCode.PUSH2] = 1 << 0,
+            [OpCode.PUSH3] = 1 << 0,
+            [OpCode.PUSH4] = 1 << 0,
+            [OpCode.PUSH5] = 1 << 0,
+            [OpCode.PUSH6] = 1 << 0,
+            [OpCode.PUSH7] = 1 << 0,
+            [OpCode.PUSH8] = 1 << 0,
+            [OpCode.PUSH9] = 1 << 0,
+            [OpCode.PUSH10] = 1 << 0,
+            [OpCode.PUSH11] = 1 << 0,
+            [OpCode.PUSH12] = 1 << 0,
+            [OpCode.PUSH13] = 1 << 0,
+            [OpCode.PUSH14] = 1 << 0,
+            [OpCode.PUSH15] = 1 << 0,
+            [OpCode.PUSH16] = 1 << 0,
+            [OpCode.NOP] = 1 << 0,
+            [OpCode.JMP] = 1 << 1,
+            [OpCode.JMP_L] = 1 << 1,
+            [OpCode.JMPIF] = 1 << 1,
+            [OpCode.JMPIF_L] = 1 << 1,
+            [OpCode.JMPIFNOT] = 1 << 1,
+            [OpCode.JMPIFNOT_L] = 1 << 1,
+            [OpCode.JMPEQ] = 1 << 1,
+            [OpCode.JMPEQ_L] = 1 << 1,
+            [OpCode.JMPNE] = 1 << 1,
+            [OpCode.JMPNE_L] = 1 << 1,
+            [OpCode.JMPGT] = 1 << 1,
+            [OpCode.JMPGT_L] = 1 << 1,
+            [OpCode.JMPGE] = 1 << 1,
+            [OpCode.JMPGE_L] = 1 << 1,
+            [OpCode.JMPLT] = 1 << 1,
+            [OpCode.JMPLT_L] = 1 << 1,
+            [OpCode.JMPLE] = 1 << 1,
+            [OpCode.JMPLE_L] = 1 << 1,
+            [OpCode.CALL] = 1 << 9,
+            [OpCode.CALL_L] = 1 << 9,
+            [OpCode.CALLA] = 1 << 9,
+            [OpCode.CALLT] = 1 << 15,
+            [OpCode.ABORT] = 0,
+            [OpCode.ASSERT] = 1 << 0,
+            [OpCode.THROW] = 1 << 9,
+            [OpCode.TRY] = 1 << 2,
+            [OpCode.TRY_L] = 1 << 2,
+            [OpCode.ENDTRY] = 1 << 2,
+            [OpCode.ENDTRY_L] = 1 << 2,
+            [OpCode.ENDFINALLY] = 1 << 2,
             [OpCode.RET] = 0,
             [OpCode.SYSCALL] = 0,
-            [OpCode.DEPTH] = 60,
-            [OpCode.DROP] = 60,
-            [OpCode.NIP] = 60,
-            [OpCode.XDROP] = 400,
-            [OpCode.CLEAR] = 400,
-            [OpCode.DUP] = 60,
-            [OpCode.OVER] = 60,
-            [OpCode.PICK] = 60,
-            [OpCode.TUCK] = 60,
-            [OpCode.SWAP] = 60,
-            [OpCode.ROT] = 60,
-            [OpCode.ROLL] = 400,
-            [OpCode.REVERSE3] = 60,
-            [OpCode.REVERSE4] = 60,
-            [OpCode.REVERSEN] = 400,
-            [OpCode.INITSSLOT] = 400,
-            [OpCode.INITSLOT] = 1600,
-            [OpCode.LDSFLD0] = 60,
-            [OpCode.LDSFLD1] = 60,
-            [OpCode.LDSFLD2] = 60,
-            [OpCode.LDSFLD3] = 60,
-            [OpCode.LDSFLD4] = 60,
-            [OpCode.LDSFLD5] = 60,
-            [OpCode.LDSFLD6] = 60,
-            [OpCode.LDSFLD] = 60,
-            [OpCode.STSFLD0] = 60,
-            [OpCode.STSFLD1] = 60,
-            [OpCode.STSFLD2] = 60,
-            [OpCode.STSFLD3] = 60,
-            [OpCode.STSFLD4] = 60,
-            [OpCode.STSFLD5] = 60,
-            [OpCode.STSFLD6] = 60,
-            [OpCode.STSFLD] = 60,
-            [OpCode.LDLOC0] = 60,
-            [OpCode.LDLOC1] = 60,
-            [OpCode.LDLOC2] = 60,
-            [OpCode.LDLOC3] = 60,
-            [OpCode.LDLOC4] = 60,
-            [OpCode.LDLOC5] = 60,
-            [OpCode.LDLOC6] = 60,
-            [OpCode.LDLOC] = 60,
-            [OpCode.STLOC0] = 60,
-            [OpCode.STLOC1] = 60,
-            [OpCode.STLOC2] = 60,
-            [OpCode.STLOC3] = 60,
-            [OpCode.STLOC4] = 60,
-            [OpCode.STLOC5] = 60,
-            [OpCode.STLOC6] = 60,
-            [OpCode.STLOC] = 60,
-            [OpCode.LDARG0] = 60,
-            [OpCode.LDARG1] = 60,
-            [OpCode.LDARG2] = 60,
-            [OpCode.LDARG3] = 60,
-            [OpCode.LDARG4] = 60,
-            [OpCode.LDARG5] = 60,
-            [OpCode.LDARG6] = 60,
-            [OpCode.LDARG] = 60,
-            [OpCode.STARG0] = 60,
-            [OpCode.STARG1] = 60,
-            [OpCode.STARG2] = 60,
-            [OpCode.STARG3] = 60,
-            [OpCode.STARG4] = 60,
-            [OpCode.STARG5] = 60,
-            [OpCode.STARG6] = 60,
-            [OpCode.STARG] = 60,
-            [OpCode.NEWBUFFER] = 80000,
-            [OpCode.MEMCPY] = 80000,
-            [OpCode.CAT] = 80000,
-            [OpCode.SUBSTR] = 80000,
-            [OpCode.LEFT] = 80000,
-            [OpCode.RIGHT] = 80000,
-            [OpCode.INVERT] = 100,
-            [OpCode.AND] = 200,
-            [OpCode.OR] = 200,
-            [OpCode.XOR] = 200,
-            [OpCode.EQUAL] = 1000,
-            [OpCode.NOTEQUAL] = 1000,
-            [OpCode.SIGN] = 100,
-            [OpCode.ABS] = 100,
-            [OpCode.NEGATE] = 100,
-            [OpCode.INC] = 100,
-            [OpCode.DEC] = 100,
-            [OpCode.ADD] = 200,
-            [OpCode.SUB] = 200,
-            [OpCode.MUL] = 300,
-            [OpCode.DIV] = 300,
-            [OpCode.MOD] = 300,
-            [OpCode.SHL] = 300,
-            [OpCode.SHR] = 300,
-            [OpCode.NOT] = 100,
-            [OpCode.BOOLAND] = 200,
-            [OpCode.BOOLOR] = 200,
-            [OpCode.NZ] = 100,
-            [OpCode.NUMEQUAL] = 200,
-            [OpCode.NUMNOTEQUAL] = 200,
-            [OpCode.LT] = 200,
-            [OpCode.LE] = 200,
-            [OpCode.GT] = 200,
-            [OpCode.GE] = 200,
-            [OpCode.MIN] = 200,
-            [OpCode.MAX] = 200,
-            [OpCode.WITHIN] = 200,
-            [OpCode.PACK] = 15000,
-            [OpCode.UNPACK] = 15000,
-            [OpCode.NEWARRAY0] = 400,
-            [OpCode.NEWARRAY] = 15000,
-            [OpCode.NEWARRAY_T] = 15000,
-            [OpCode.NEWSTRUCT0] = 400,
-            [OpCode.NEWSTRUCT] = 15000,
-            [OpCode.NEWMAP] = 200,
-            [OpCode.SIZE] = 150,
-            [OpCode.HASKEY] = 270000,
-            [OpCode.KEYS] = 500,
-            [OpCode.VALUES] = 270000,
-            [OpCode.PICKITEM] = 270000,
-            [OpCode.APPEND] = 270000,
-            [OpCode.SETITEM] = 270000,
-            [OpCode.REVERSEITEMS] = 270000,
-            [OpCode.REMOVE] = 500,
-            [OpCode.CLEARITEMS] = 400,
-            [OpCode.ISNULL] = 60,
-            [OpCode.ISTYPE] = 60,
-            [OpCode.CONVERT] = 80000,
+            [OpCode.DEPTH] = 1 << 1,
+            [OpCode.DROP] = 1 << 1,
+            [OpCode.NIP] = 1 << 1,
+            [OpCode.XDROP] = 1 << 4,
+            [OpCode.CLEAR] = 1 << 4,
+            [OpCode.DUP] = 1 << 1,
+            [OpCode.OVER] = 1 << 1,
+            [OpCode.PICK] = 1 << 1,
+            [OpCode.TUCK] = 1 << 1,
+            [OpCode.SWAP] = 1 << 1,
+            [OpCode.ROT] = 1 << 1,
+            [OpCode.ROLL] = 1 << 4,
+            [OpCode.REVERSE3] = 1 << 1,
+            [OpCode.REVERSE4] = 1 << 1,
+            [OpCode.REVERSEN] = 1 << 4,
+            [OpCode.INITSSLOT] = 1 << 4,
+            [OpCode.INITSLOT] = 1 << 6,
+            [OpCode.LDSFLD0] = 1 << 1,
+            [OpCode.LDSFLD1] = 1 << 1,
+            [OpCode.LDSFLD2] = 1 << 1,
+            [OpCode.LDSFLD3] = 1 << 1,
+            [OpCode.LDSFLD4] = 1 << 1,
+            [OpCode.LDSFLD5] = 1 << 1,
+            [OpCode.LDSFLD6] = 1 << 1,
+            [OpCode.LDSFLD] = 1 << 1,
+            [OpCode.STSFLD0] = 1 << 1,
+            [OpCode.STSFLD1] = 1 << 1,
+            [OpCode.STSFLD2] = 1 << 1,
+            [OpCode.STSFLD3] = 1 << 1,
+            [OpCode.STSFLD4] = 1 << 1,
+            [OpCode.STSFLD5] = 1 << 1,
+            [OpCode.STSFLD6] = 1 << 1,
+            [OpCode.STSFLD] = 1 << 1,
+            [OpCode.LDLOC0] = 1 << 1,
+            [OpCode.LDLOC1] = 1 << 1,
+            [OpCode.LDLOC2] = 1 << 1,
+            [OpCode.LDLOC3] = 1 << 1,
+            [OpCode.LDLOC4] = 1 << 1,
+            [OpCode.LDLOC5] = 1 << 1,
+            [OpCode.LDLOC6] = 1 << 1,
+            [OpCode.LDLOC] = 1 << 1,
+            [OpCode.STLOC0] = 1 << 1,
+            [OpCode.STLOC1] = 1 << 1,
+            [OpCode.STLOC2] = 1 << 1,
+            [OpCode.STLOC3] = 1 << 1,
+            [OpCode.STLOC4] = 1 << 1,
+            [OpCode.STLOC5] = 1 << 1,
+            [OpCode.STLOC6] = 1 << 1,
+            [OpCode.STLOC] = 1 << 1,
+            [OpCode.LDARG0] = 1 << 1,
+            [OpCode.LDARG1] = 1 << 1,
+            [OpCode.LDARG2] = 1 << 1,
+            [OpCode.LDARG3] = 1 << 1,
+            [OpCode.LDARG4] = 1 << 1,
+            [OpCode.LDARG5] = 1 << 1,
+            [OpCode.LDARG6] = 1 << 1,
+            [OpCode.LDARG] = 1 << 1,
+            [OpCode.STARG0] = 1 << 1,
+            [OpCode.STARG1] = 1 << 1,
+            [OpCode.STARG2] = 1 << 1,
+            [OpCode.STARG3] = 1 << 1,
+            [OpCode.STARG4] = 1 << 1,
+            [OpCode.STARG5] = 1 << 1,
+            [OpCode.STARG6] = 1 << 1,
+            [OpCode.STARG] = 1 << 1,
+            [OpCode.NEWBUFFER] = 1 << 8,
+            [OpCode.MEMCPY] = 1 << 11,
+            [OpCode.CAT] = 1 << 11,
+            [OpCode.SUBSTR] = 1 << 11,
+            [OpCode.LEFT] = 1 << 11,
+            [OpCode.RIGHT] = 1 << 11,
+            [OpCode.INVERT] = 1 << 2,
+            [OpCode.AND] = 1 << 3,
+            [OpCode.OR] = 1 << 3,
+            [OpCode.XOR] = 1 << 3,
+            [OpCode.EQUAL] = 1 << 5,
+            [OpCode.NOTEQUAL] = 1 << 5,
+            [OpCode.SIGN] = 1 << 2,
+            [OpCode.ABS] = 1 << 2,
+            [OpCode.NEGATE] = 1 << 2,
+            [OpCode.INC] = 1 << 2,
+            [OpCode.DEC] = 1 << 2,
+            [OpCode.ADD] = 1 << 3,
+            [OpCode.SUB] = 1 << 3,
+            [OpCode.MUL] = 1 << 3,
+            [OpCode.DIV] = 1 << 3,
+            [OpCode.MOD] = 1 << 3,
+            [OpCode.SHL] = 1 << 3,
+            [OpCode.SHR] = 1 << 3,
+            [OpCode.NOT] = 1 << 2,
+            [OpCode.BOOLAND] = 1 << 3,
+            [OpCode.BOOLOR] = 1 << 3,
+            [OpCode.NZ] = 1 << 2,
+            [OpCode.NUMEQUAL] = 1 << 3,
+            [OpCode.NUMNOTEQUAL] = 1 << 3,
+            [OpCode.LT] = 1 << 3,
+            [OpCode.LE] = 1 << 3,
+            [OpCode.GT] = 1 << 3,
+            [OpCode.GE] = 1 << 3,
+            [OpCode.MIN] = 1 << 3,
+            [OpCode.MAX] = 1 << 3,
+            [OpCode.WITHIN] = 1 << 3,
+            [OpCode.PACK] = 1 << 9,
+            [OpCode.UNPACK] = 1 << 9,
+            [OpCode.NEWARRAY0] = 1 << 4,
+            [OpCode.NEWARRAY] = 1 << 9,
+            [OpCode.NEWARRAY_T] = 1 << 9,
+            [OpCode.NEWSTRUCT0] = 1 << 4,
+            [OpCode.NEWSTRUCT] = 1 << 9,
+            [OpCode.NEWMAP] = 1 << 3,
+            [OpCode.SIZE] = 1 << 2,
+            [OpCode.HASKEY] = 1 << 6,
+            [OpCode.KEYS] = 1 << 4,
+            [OpCode.VALUES] = 1 << 13,
+            [OpCode.PICKITEM] = 1 << 6,
+            [OpCode.APPEND] = 1 << 13,
+            [OpCode.SETITEM] = 1 << 13,
+            [OpCode.REVERSEITEMS] = 1 << 13,
+            [OpCode.REMOVE] = 1 << 4,
+            [OpCode.CLEARITEMS] = 1 << 4,
+            [OpCode.ISNULL] = 1 << 1,
+            [OpCode.ISTYPE] = 1 << 1,
+            [OpCode.CONVERT] = 1 << 11,
         };
     }
 }
