@@ -2,10 +2,9 @@ using Neo.Cryptography.ECC;
 using Neo.IO;
 using Neo.Network.P2P.Payloads;
 using Neo.SmartContract.Native;
-using Neo.SmartContract.Native.Oracle;
-using Neo.VM.Types;
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using Array = Neo.VM.Types.Array;
 
@@ -16,65 +15,19 @@ namespace Neo.SmartContract
         public const int MaxEventName = 32;
         public const int MaxNotificationSize = 1024;
 
-        public static readonly InteropDescriptor System_Runtime_Platform = Register("System.Runtime.Platform", nameof(GetPlatform), 0_00000250, CallFlags.None, true);
-        public static readonly InteropDescriptor System_Runtime_GetTrigger = Register("System.Runtime.GetTrigger", nameof(Trigger), 0_00000250, CallFlags.None, true);
-        public static readonly InteropDescriptor System_Runtime_GetTime = Register("System.Runtime.GetTime", nameof(GetTime), 0_00000250, CallFlags.AllowStates, true);
-        public static readonly InteropDescriptor System_Runtime_GetScriptContainer = Register("System.Runtime.GetScriptContainer", nameof(GetScriptContainer), 0_00000250, CallFlags.None, true);
-        public static readonly InteropDescriptor System_Runtime_GetExecutingScriptHash = Register("System.Runtime.GetExecutingScriptHash", nameof(CurrentScriptHash), 0_00000400, CallFlags.None, true);
-        public static readonly InteropDescriptor System_Runtime_GetCallingScriptHash = Register("System.Runtime.GetCallingScriptHash", nameof(CallingScriptHash), 0_00000400, CallFlags.None, true);
-        public static readonly InteropDescriptor System_Runtime_GetEntryScriptHash = Register("System.Runtime.GetEntryScriptHash", nameof(EntryScriptHash), 0_00000400, CallFlags.None, true);
-        public static readonly InteropDescriptor System_Runtime_CheckWitness = Register("System.Runtime.CheckWitness", nameof(CheckWitness), 0_00030000, CallFlags.None, true);
-        public static readonly InteropDescriptor System_Runtime_GetInvocationCounter = Register("System.Runtime.GetInvocationCounter", nameof(GetInvocationCounter), 0_00000400, CallFlags.None, true);
-        public static readonly InteropDescriptor System_Runtime_Log = Register("System.Runtime.Log", nameof(RuntimeLog), 0_01000000, CallFlags.AllowNotify, false);
-        public static readonly InteropDescriptor System_Runtime_Notify = Register("System.Runtime.Notify", nameof(RuntimeNotify), 0_01000000, CallFlags.AllowNotify, false);
-        public static readonly InteropDescriptor System_Runtime_GetNotifications = Register("System.Runtime.GetNotifications", nameof(GetNotifications), 0_00010000, CallFlags.None, true);
-        public static readonly InteropDescriptor System_Runtime_GasLeft = Register("System.Runtime.GasLeft", nameof(GasLeft), 0_00000400, CallFlags.None, true);
-
-        private static bool CheckItemForNotification(StackItem state)
-        {
-            int size = 0;
-            List<StackItem> items_checked = new List<StackItem>();
-            Queue<StackItem> items_unchecked = new Queue<StackItem>();
-            while (true)
-            {
-                switch (state)
-                {
-                    case Struct array:
-                        foreach (StackItem item in array)
-                            items_unchecked.Enqueue(item);
-                        break;
-                    case Array array:
-                        if (items_checked.All(p => !ReferenceEquals(p, array)))
-                        {
-                            items_checked.Add(array);
-                            foreach (StackItem item in array)
-                                items_unchecked.Enqueue(item);
-                        }
-                        break;
-                    case PrimitiveType primitive:
-                        size += primitive.Size;
-                        break;
-                    case Null _:
-                        break;
-                    case InteropInterface _:
-                        return false;
-                    case Map map:
-                        if (items_checked.All(p => !ReferenceEquals(p, map)))
-                        {
-                            items_checked.Add(map);
-                            foreach (var pair in map)
-                            {
-                                size += pair.Key.Size;
-                                items_unchecked.Enqueue(pair.Value);
-                            }
-                        }
-                        break;
-                }
-                if (size > MaxNotificationSize) return false;
-                if (items_unchecked.Count == 0) return true;
-                state = items_unchecked.Dequeue();
-            }
-        }
+        public static readonly InteropDescriptor System_Runtime_Platform = Register("System.Runtime.Platform", nameof(GetPlatform), 1 << 3, CallFlags.None);
+        public static readonly InteropDescriptor System_Runtime_GetTrigger = Register("System.Runtime.GetTrigger", nameof(Trigger), 1 << 3, CallFlags.None);
+        public static readonly InteropDescriptor System_Runtime_GetTime = Register("System.Runtime.GetTime", nameof(GetTime), 1 << 3, CallFlags.None);
+        public static readonly InteropDescriptor System_Runtime_GetScriptContainer = Register("System.Runtime.GetScriptContainer", nameof(GetScriptContainer), 1 << 3, CallFlags.None);
+        public static readonly InteropDescriptor System_Runtime_GetExecutingScriptHash = Register("System.Runtime.GetExecutingScriptHash", nameof(CurrentScriptHash), 1 << 4, CallFlags.None);
+        public static readonly InteropDescriptor System_Runtime_GetCallingScriptHash = Register("System.Runtime.GetCallingScriptHash", nameof(CallingScriptHash), 1 << 4, CallFlags.None);
+        public static readonly InteropDescriptor System_Runtime_GetEntryScriptHash = Register("System.Runtime.GetEntryScriptHash", nameof(EntryScriptHash), 1 << 4, CallFlags.None);
+        public static readonly InteropDescriptor System_Runtime_CheckWitness = Register("System.Runtime.CheckWitness", nameof(CheckWitness), 1 << 10, CallFlags.None);
+        public static readonly InteropDescriptor System_Runtime_GetInvocationCounter = Register("System.Runtime.GetInvocationCounter", nameof(GetInvocationCounter), 1 << 4, CallFlags.None);
+        public static readonly InteropDescriptor System_Runtime_Log = Register("System.Runtime.Log", nameof(RuntimeLog), 1 << 15, CallFlags.AllowNotify);
+        public static readonly InteropDescriptor System_Runtime_Notify = Register("System.Runtime.Notify", nameof(RuntimeNotify), 1 << 15, CallFlags.AllowNotify);
+        public static readonly InteropDescriptor System_Runtime_GetNotifications = Register("System.Runtime.GetNotifications", nameof(GetNotifications), 1 << 8, CallFlags.None);
+        public static readonly InteropDescriptor System_Runtime_GasLeft = Register("System.Runtime.GasLeft", nameof(GasLeft), 1 << 4, CallFlags.None);
 
         protected internal string GetPlatform()
         {
@@ -136,10 +89,10 @@ namespace Neo.SmartContract
                 {
                     // Check allow state callflag
 
-                    if (!CurrentContext.GetState<ExecutionContextState>().CallFlags.HasFlag(CallFlags.AllowStates))
+                    if (!CurrentContext.GetState<ExecutionContextState>().CallFlags.HasFlag(CallFlags.ReadStates))
                         throw new InvalidOperationException($"Cannot call this SYSCALL without the flag AllowStates.");
 
-                    var contract = Snapshot.Contracts[CallingScriptHash];
+                    var contract = NativeContract.ContractManagement.GetContract(Snapshot, CallingScriptHash);
                     // check if current group is the required one
                     if (contract.Manifest.Groups.Select(p => p.PubKey).Intersect(signer.AllowedGroups).Any())
                         return true;
@@ -149,7 +102,7 @@ namespace Neo.SmartContract
 
             // Check allow state callflag
 
-            if (!CurrentContext.GetState<ExecutionContextState>().CallFlags.HasFlag(CallFlags.AllowStates))
+            if (!CurrentContext.GetState<ExecutionContextState>().CallFlags.HasFlag(CallFlags.ReadStates))
                 throw new InvalidOperationException($"Cannot call this SYSCALL without the flag AllowStates.");
 
             // only for non-Transaction types (Block, etc)
@@ -177,7 +130,11 @@ namespace Neo.SmartContract
         protected internal void RuntimeNotify(byte[] eventName, Array state)
         {
             if (eventName.Length > MaxEventName) throw new ArgumentException();
-            if (!CheckItemForNotification(state)) throw new ArgumentException();
+            using (MemoryStream ms = new MemoryStream(MaxNotificationSize))
+            using (BinaryWriter writer = new BinaryWriter(ms))
+            {
+                BinarySerializer.Serialize(writer, state, MaxNotificationSize);
+            }
             SendNotification(CurrentScriptHash, Utility.StrictUTF8.GetString(eventName), state);
         }
 
