@@ -7,12 +7,13 @@ namespace Neo.SmartContract
     {
         None = 0,
 
-        AllowStates = 0b00000001,
-        AllowModifyStates = 0b00000010,
+        ReadStates = 0b00000001,
+        WriteStates = 0b00000010,
         AllowCall = 0b00000100,
         AllowNotify = 0b00001000,
 
-        ReadOnly = AllowStates | AllowCall | AllowNotify,
-        All = AllowStates | AllowModifyStates | AllowCall | AllowNotify
+        States = ReadStates | WriteStates,
+        ReadOnly = ReadStates | AllowCall,
+        All = States | AllowCall | AllowNotify
     }
 }
