@@ -1,4 +1,3 @@
-using Neo.Ledger;
 using Neo.SmartContract.Manifest;
 using System;
 using System.Linq;
@@ -14,7 +13,7 @@ namespace Neo.SmartContract
             if (contract is null) throw new ArgumentNullException(nameof(contract));
 
             Script = null;
-            ScriptHash = contract.ScriptHash;
+            ScriptHash = contract.Hash;
             ContractMethodDescriptor descriptor = contract.Manifest.Abi.GetMethod("verify");
             if (descriptor is null) throw new NotSupportedException("The smart contract haven't got verify method.");
 
