@@ -1,3 +1,4 @@
+using Neo.VM;
 using Neo.VM.Types;
 using System.Collections.Generic;
 
@@ -6,20 +7,17 @@ namespace Neo.SmartContract.Iterators
     internal class MapWrapper : IIterator
     {
         private readonly IEnumerator<KeyValuePair<PrimitiveType, StackItem>> enumerator;
+        private readonly ReferenceCounter referenceCounter;
 
-        public MapWrapper(IEnumerable<KeyValuePair<PrimitiveType, StackItem>> map)
+        public MapWrapper(IEnumerable<KeyValuePair<PrimitiveType, StackItem>> map, ReferenceCounter referenceCounter)
         {
             this.enumerator = map.GetEnumerator();
+            this.referenceCounter = referenceCounter;
         }
 
         public void Dispose()
         {
             enumerator.Dispose();
-        }
-
-        public PrimitiveType Key()
-        {
-            return enumerator.Current.Key;
         }
 
         public bool Next()
@@ -29,7 +27,7 @@ namespace Neo.SmartContract.Iterators
 
         public StackItem Value()
         {
-            return enumerator.Current.Value;
+            return new Struct(referenceCounter) { enumerator.Current.Key, enumerator.Current.Value };
         }
     }
 }
