@@ -15,7 +15,7 @@ namespace Neo.SmartContract.Manifest
         /// <summary>
         /// Max length for a valid Contract Manifest
         /// </summary>
-        public const int MaxLength = 4096;
+        public const int MaxLength = ushort.MaxValue;
 
         /// <summary>
         /// Serialized size
@@ -30,9 +30,9 @@ namespace Neo.SmartContract.Manifest
         }
 
         /// <summary>
-        /// Contract hash
+        /// Contract name
         /// </summary>
-        public UInt160 Hash => Abi.Hash;
+        public string Name { get; set; }
 
         /// <summary>
         /// A group represents a set of mutually trusted contracts. A contract will trust and allow any contract in the same group to invoke it, and the user interface will not give any warnings.
@@ -61,26 +61,9 @@ namespace Neo.SmartContract.Manifest
         public WildcardContainer<UInt160> Trusts { get; set; }
 
         /// <summary>
-        /// The safemethods field is an array containing a set of method names. It can also be assigned with a wildcard *. If it is a wildcard *, then it means that all methods of the contract are safe.
-        /// If a method is marked as safe, the user interface will not give any warnings when it is called by any other contract.
-        /// </summary>
-        public WildcardContainer<string> SafeMethods { get; set; }
-
-        /// <summary>
         /// Custom user data
         /// </summary>
         public JObject Extra { get; set; }
-
-        /// <summary>
-        /// Return true if is allowed
-        /// </summary>
-        /// <param name="manifest">Manifest</param>
-        /// <param name="method">Method</param>
-        /// <returns>Return true or false</returns>
-        public bool CanCall(ContractManifest manifest, string method)
-        {
-            return Permissions.Any(u => u.IsAllowed(manifest, method));
-        }
 
         /// <summary>
         /// Parse ContractManifest from json
@@ -110,12 +93,12 @@ namespace Neo.SmartContract.Manifest
         {
             return new JObject
             {
+                ["name"] = Name,
                 ["groups"] = Groups.Select(u => u.ToJson()).ToArray(),
                 ["supportedstandards"] = SupportedStandards.Select(u => new JString(u)).ToArray(),
                 ["abi"] = Abi.ToJson(),
                 ["permissions"] = Permissions.Select(p => p.ToJson()).ToArray(),
                 ["trusts"] = Trusts.ToJson(),
-                ["safemethods"] = SafeMethods.ToJson(),
                 ["extra"] = Extra
             };
         }
@@ -128,12 +111,12 @@ namespace Neo.SmartContract.Manifest
         {
             return new ContractManifest
             {
+                Name = Name,
                 Groups = Groups.Select(p => p.Clone()).ToArray(),
                 SupportedStandards = SupportedStandards[..],
                 Abi = Abi.Clone(),
                 Permissions = Permissions.Select(p => p.Clone()).ToArray(),
                 Trusts = Trusts,
-                SafeMethods = SafeMethods,
                 Extra = Extra?.Clone()
             };
         }
@@ -156,12 +139,12 @@ namespace Neo.SmartContract.Manifest
 
         private void DeserializeFromJson(JObject json)
         {
+            Name = json["name"].AsString();
             Groups = ((JArray)json["groups"]).Select(u => ContractGroup.FromJson(u)).ToArray();
             SupportedStandards = ((JArray)json["supportedstandards"]).Select(u => u.AsString()).ToArray();
             Abi = ContractAbi.FromJson(json["abi"]);
             Permissions = ((JArray)json["permissions"]).Select(u => ContractPermission.FromJson(u)).ToArray();
             Trusts = WildcardContainer<UInt160>.FromJson(json["trusts"], u => UInt160.Parse(u.AsString()));
-            SafeMethods = WildcardContainer<string>.FromJson(json["safemethods"], u => u.AsString());
             Extra = json["extra"];
         }
 
@@ -171,7 +154,6 @@ namespace Neo.SmartContract.Manifest
         /// <returns>Return true or false</returns>
         public bool IsValid(UInt160 hash)
         {
-            if (!Abi.Hash.Equals(hash)) return false;
             return Groups.All(u => u.IsValid(hash));
         }
     }
