@@ -64,20 +64,20 @@ namespace Neo.SmartContract.Manifest
         /// <summary>
         /// Return true if is allowed
         /// </summary>
-        /// <param name="manifest">The manifest of which contract we are calling</param>
-        /// <param name="method">Method</param>
+        /// <param name="targetContract">The contract that we are calling</param>
+        /// <param name="targetMethod">The method that we are calling</param>
         /// <returns>Return true or false</returns>
-        public bool IsAllowed(ContractManifest manifest, string method)
+        public bool IsAllowed(ContractState targetContract, string targetMethod)
         {
             if (Contract.IsHash)
             {
-                if (!Contract.Hash.Equals(manifest.Hash)) return false;
+                if (!Contract.Hash.Equals(targetContract.Hash)) return false;
             }
             else if (Contract.IsGroup)
             {
-                if (manifest.Groups.All(p => !p.PubKey.Equals(Contract.Group))) return false;
+                if (targetContract.Manifest.Groups.All(p => !p.PubKey.Equals(Contract.Group))) return false;
             }
-            return Methods.IsWildcard || Methods.Contains(method);
+            return Methods.IsWildcard || Methods.Contains(targetMethod);
         }
     }
 }
