@@ -11,20 +11,26 @@ namespace Neo.SmartContract.Native
 {
     public sealed class PolicyContract : NativeContract
     {
-        public override string Name => "Policy";
         public override int Id => -3;
+
+        public const uint DefaultExecFeeFactor = 30;
+        public const uint DefaultStoragePrice = 100000;
+        private const uint MaxExecFeeFactor = 1000;
+        private const uint MaxStoragePrice = 10000000;
 
         private const byte Prefix_MaxTransactionsPerBlock = 23;
         private const byte Prefix_FeePerByte = 10;
         private const byte Prefix_BlockedAccount = 15;
         private const byte Prefix_MaxBlockSize = 12;
         private const byte Prefix_MaxBlockSystemFee = 17;
+        private const byte Prefix_ExecFeeFactor = 18;
+        private const byte Prefix_StoragePrice = 19;
 
         internal PolicyContract()
         {
         }
 
-        [ContractMethod(0_01000000, CallFlags.AllowStates)]
+        [ContractMethod(0_01000000, CallFlags.ReadStates)]
         public uint GetMaxTransactionsPerBlock(StoreView snapshot)
         {
             StorageItem item = snapshot.Storages.TryGet(CreateStorageKey(Prefix_MaxTransactionsPerBlock));
@@ -32,7 +38,7 @@ namespace Neo.SmartContract.Native
             return (uint)(BigInteger)item;
         }
 
-        [ContractMethod(0_01000000, CallFlags.AllowStates)]
+        [ContractMethod(0_01000000, CallFlags.ReadStates)]
         public uint GetMaxBlockSize(StoreView snapshot)
         {
             StorageItem item = snapshot.Storages.TryGet(CreateStorageKey(Prefix_MaxBlockSize));
@@ -40,7 +46,7 @@ namespace Neo.SmartContract.Native
             return (uint)(BigInteger)item;
         }
 
-        [ContractMethod(0_01000000, CallFlags.AllowStates)]
+        [ContractMethod(0_01000000, CallFlags.ReadStates)]
         public long GetMaxBlockSystemFee(StoreView snapshot)
         {
             StorageItem item = snapshot.Storages.TryGet(CreateStorageKey(Prefix_MaxBlockSystemFee));
@@ -48,7 +54,7 @@ namespace Neo.SmartContract.Native
             return (long)(BigInteger)item;
         }
 
-        [ContractMethod(0_01000000, CallFlags.AllowStates)]
+        [ContractMethod(0_01000000, CallFlags.ReadStates)]
         public long GetFeePerByte(StoreView snapshot)
         {
             StorageItem item = snapshot.Storages.TryGet(CreateStorageKey(Prefix_FeePerByte));
@@ -56,13 +62,29 @@ namespace Neo.SmartContract.Native
             return (long)(BigInteger)item;
         }
 
-        [ContractMethod(0_01000000, CallFlags.AllowStates)]
+        [ContractMethod(0_01000000, CallFlags.ReadStates)]
+        public uint GetExecFeeFactor(StoreView snapshot)
+        {
+            StorageItem item = snapshot.Storages.TryGet(CreateStorageKey(Prefix_ExecFeeFactor));
+            if (item is null) return DefaultExecFeeFactor;
+            return (uint)(BigInteger)item;
+        }
+
+        [ContractMethod(0_01000000, CallFlags.ReadStates)]
+        public uint GetStoragePrice(StoreView snapshot)
+        {
+            StorageItem item = snapshot.Storages.TryGet(CreateStorageKey(Prefix_StoragePrice));
+            if (item is null) return DefaultStoragePrice;
+            return (uint)(BigInteger)item;
+        }
+
+        [ContractMethod(0_01000000, CallFlags.ReadStates)]
         public bool IsBlocked(StoreView snapshot, UInt160 account)
         {
             return snapshot.Storages.Contains(CreateStorageKey(Prefix_BlockedAccount).Add(account));
         }
 
-        [ContractMethod(0_03000000, CallFlags.AllowModifyStates)]
+        [ContractMethod(0_03000000, CallFlags.WriteStates)]
         private bool SetMaxBlockSize(ApplicationEngine engine, uint value)
         {
             if (value > Message.PayloadMaxSize) throw new ArgumentOutOfRangeException(nameof(value));
@@ -72,7 +94,7 @@ namespace Neo.SmartContract.Native
             return true;
         }
 
-        [ContractMethod(0_03000000, CallFlags.AllowModifyStates)]
+        [ContractMethod(0_03000000, CallFlags.WriteStates)]
         private bool SetMaxTransactionsPerBlock(ApplicationEngine engine, uint value)
         {
             if (value > Block.MaxTransactionsPerBlock) throw new ArgumentOutOfRangeException(nameof(value));
@@ -82,7 +104,7 @@ namespace Neo.SmartContract.Native
             return true;
         }
 
-        [ContractMethod(0_03000000, CallFlags.AllowModifyStates)]
+        [ContractMethod(0_03000000, CallFlags.WriteStates)]
         private bool SetMaxBlockSystemFee(ApplicationEngine engine, long value)
         {
             if (value <= 4007600) throw new ArgumentOutOfRangeException(nameof(value));
@@ -92,7 +114,7 @@ namespace Neo.SmartContract.Native
             return true;
         }
 
-        [ContractMethod(0_03000000, CallFlags.AllowModifyStates)]
+        [ContractMethod(0_03000000, CallFlags.WriteStates)]
         private bool SetFeePerByte(ApplicationEngine engine, long value)
         {
             if (value < 0 || value > 1_00000000) throw new ArgumentOutOfRangeException(nameof(value));
@@ -102,7 +124,27 @@ namespace Neo.SmartContract.Native
             return true;
         }
 
-        [ContractMethod(0_03000000, CallFlags.AllowModifyStates)]
+        [ContractMethod(0_03000000, CallFlags.WriteStates)]
+        private bool SetExecFeeFactor(ApplicationEngine engine, uint value)
+        {
+            if (value == 0 || value > MaxExecFeeFactor) throw new ArgumentOutOfRangeException(nameof(value));
+            if (!CheckCommittee(engine)) return false;
+            StorageItem storage = engine.Snapshot.Storages.GetAndChange(CreateStorageKey(Prefix_ExecFeeFactor), () => new StorageItem());
+            storage.Set(value);
+            return true;
+        }
+
+        [ContractMethod(0_03000000, CallFlags.WriteStates)]
+        private bool SetStoragePrice(ApplicationEngine engine, uint value)
+        {
+            if (value == 0 || value > MaxStoragePrice) throw new ArgumentOutOfRangeException(nameof(value));
+            if (!CheckCommittee(engine)) return false;
+            StorageItem storage = engine.Snapshot.Storages.GetAndChange(CreateStorageKey(Prefix_StoragePrice), () => new StorageItem());
+            storage.Set(value);
+            return true;
+        }
+
+        [ContractMethod(0_03000000, CallFlags.WriteStates)]
         private bool BlockAccount(ApplicationEngine engine, UInt160 account)
         {
             if (!CheckCommittee(engine)) return false;
@@ -114,7 +156,7 @@ namespace Neo.SmartContract.Native
             return true;
         }
 
-        [ContractMethod(0_03000000, CallFlags.AllowModifyStates)]
+        [ContractMethod(0_03000000, CallFlags.WriteStates)]
         private bool UnblockAccount(ApplicationEngine engine, UInt160 account)
         {
             if (!CheckCommittee(engine)) return false;
