@@ -1,23 +1,29 @@
 using Neo.Cryptography;
 using Neo.IO;
+using Neo.IO.Json;
 using System;
 using System.IO;
+using System.Linq;
 
 namespace Neo.SmartContract
 {
     /// <summary>
-    /// +------------+-----------+------------------------------------------------------------+
-    /// |   Field    |  Length   |                          Comment                           |
-    /// +------------+-----------+------------------------------------------------------------+
-    /// | Magic      | 4 bytes   | Magic header                                               |
-    /// | Compiler   | 32 bytes  | Compiler used                                              |
-    /// | Version    | 16 bytes  | Compiler version (Mayor, Minor, Build, Version)            |
-    /// | ScriptHash | 20 bytes  | ScriptHash for the script                                  |
-    /// +------------+-----------+------------------------------------------------------------+
-    /// | Checksum   | 4 bytes   | First four bytes of double SHA256 hash                     |
-    /// +------------+-----------+------------------------------------------------------------+
-    /// | Script     | Var bytes | Var bytes for the payload                                  |
-    /// +------------+-----------+------------------------------------------------------------+
+    /// ┌───────────────────────────────────────────────────────────────────────┐
+    /// │                    NEO Executable Format 3 (NEF3)                     │
+    /// ├──────────┬───────────────┬────────────────────────────────────────────┤
+    /// │  Field   │     Type      │                  Comment                   │
+    /// ├──────────┼───────────────┼────────────────────────────────────────────┤
+    /// │ Magic    │ uint32        │ Magic header                               │
+    /// │ Compiler │ byte[32]      │ Compiler used                              │
+    /// │ Version  │ byte[32]      │ Compiler version                           │
+    /// ├──────────┼───────────────┼────────────────────────────────────────────┤
+    /// │ Reserve  │ byte[2]       │ Reserved for future extensions. Must be 0. │
+    /// │ Tokens   │ MethodToken[] │ Method tokens.                             │
+    /// │ Reserve  │ byte[2]       │ Reserved for future extensions. Must be 0. │
+    /// │ Script   │ byte[]        │ Var bytes for the payload                  │
+    /// ├──────────┼───────────────┼────────────────────────────────────────────┤
+    /// │ Checksum │ uint32        │ First four bytes of double SHA256 hash     │
+    /// └──────────┴───────────────┴────────────────────────────────────────────┘
     /// </summary>
     public class NefFile : ISerializable
     {
@@ -34,39 +40,45 @@ namespace Neo.SmartContract
         /// <summary>
         /// Version
         /// </summary>
-        public Version Version { get; set; }
+        public string Version { get; set; }
 
         /// <summary>
-        /// Script Hash
+        /// Method tokens
         /// </summary>
-        public UInt160 ScriptHash { get; set; }
-
-        /// <summary>
-        /// Checksum
-        /// </summary>
-        public uint CheckSum { get; set; }
+        public MethodToken[] Tokens { get; set; }
 
         /// <summary>
         /// Script
         /// </summary>
         public byte[] Script { get; set; }
 
+        /// <summary>
+        /// Checksum
+        /// </summary>
+        public uint CheckSum { get; set; }
+
+        public const int MaxScriptLength = 512 * 1024;
+
         private const int HeaderSize =
             sizeof(uint) +      // Magic
-            32 +                // Compiler
-            (sizeof(int) * 4) + // Version
-            UInt160.Length;     // ScriptHash
+            (32 * 2);           // Compiler+Version
 
         public int Size =>
-            HeaderSize +        // Header
-            sizeof(uint) +      // Checksum
-            Script.GetVarSize();// Script
+            HeaderSize +            // Header
+            2 +                     // Reserve
+            Tokens.GetVarSize() +   // Tokens
+            2 +                     // Reserve
+            Script.GetVarSize() +   // Script
+            sizeof(uint);           // Checksum
 
         public void Serialize(BinaryWriter writer)
         {
             SerializeHeader(writer);
-            writer.Write(CheckSum);
+            writer.Write((short)0);
+            writer.Write(Tokens);
+            writer.Write((short)0);
             writer.WriteVarBytes(Script ?? Array.Empty<byte>());
+            writer.Write(CheckSum);
         }
 
         private void SerializeHeader(BinaryWriter writer)
@@ -75,37 +87,27 @@ namespace Neo.SmartContract
             writer.WriteFixedString(Compiler, 32);
 
             // Version
-            writer.Write(Version.Major);
-            writer.Write(Version.Minor);
-            writer.Write(Version.Build);
-            writer.Write(Version.Revision);
-
-            writer.Write(ScriptHash);
+            writer.WriteFixedString(Version, 32);
         }
 
         public void Deserialize(BinaryReader reader)
         {
-            if (reader.ReadUInt32() != Magic)
-            {
-                throw new FormatException("Wrong magic");
-            }
+            if (reader.ReadUInt32() != Magic) throw new FormatException("Wrong magic");
 
             Compiler = reader.ReadFixedString(32);
-            Version = new Version(reader.ReadInt32(), reader.ReadInt32(), reader.ReadInt32(), reader.ReadInt32());
-            ScriptHash = reader.ReadSerializable<UInt160>();
+            Version = reader.ReadFixedString(32);
+
+            if (reader.ReadUInt16() != 0) throw new FormatException("Reserved bytes must be 0");
+
+            Tokens = reader.ReadSerializableArray<MethodToken>(128);
+
+            if (reader.ReadUInt16() != 0) throw new FormatException("Reserved bytes must be 0");
+
+            Script = reader.ReadVarBytes(MaxScriptLength);
+            if (Script.Length == 0) throw new ArgumentException($"Script can't be empty");
+
             CheckSum = reader.ReadUInt32();
-
-            if (CheckSum != ComputeChecksum(this))
-            {
-                throw new FormatException("CRC verification fail");
-            }
-
-            Script = reader.ReadVarBytes(1024 * 1024);
-
-            if (Script.ToScriptHash() != ScriptHash)
-            {
-                throw new FormatException("ScriptHash is different");
-            }
+            if (CheckSum != ComputeChecksum(this)) throw new FormatException("CRC verification fail");
         }
 
         /// <summary>
@@ -113,17 +115,22 @@ namespace Neo.SmartContract
         /// </summary>
         /// <param name="file">File</param>
         /// <returns>Return checksum</returns>
-        unsafe public static uint ComputeChecksum(NefFile file)
+        public static uint ComputeChecksum(NefFile file)
         {
-            Span<byte> header = stackalloc byte[HeaderSize];
-            fixed (byte* p = header)
-                using (UnmanagedMemoryStream ms = new UnmanagedMemoryStream(p, HeaderSize, HeaderSize, FileAccess.Write))
-                using (BinaryWriter wr = new BinaryWriter(ms, Utility.StrictUTF8, false))
-                {
-                    file.SerializeHeader(wr);
-                    wr.Flush();
-                }
-            return BitConverter.ToUInt32(Crypto.Hash256(header), 0);
+            return BitConverter.ToUInt32(Crypto.Hash256(file.ToArray().AsSpan(..^sizeof(int))));
+        }
+
+        public JObject ToJson()
+        {
+            return new JObject
+            {
+                ["magic"] = Magic,
+                ["compiler"] = Compiler,
+                ["version"] = Version,
+                ["tokens"] = new JArray(Tokens.Select(p => p.ToJson())),
+                ["script"] = Convert.ToBase64String(Script),
+                ["checksum"] = CheckSum
+            };
         }
     }
 }
