@@ -1,4 +1,7 @@
+using Neo.Ledger;
+using Neo.Persistence;
 using Neo.SmartContract;
+using Neo.SmartContract.Native;
 using Neo.VM;
 using System;
 
@@ -6,23 +9,29 @@ namespace Neo.Wallets
 {
     public class AssetDescriptor
     {
-        public UInt160 AssetId;
-        public string AssetName;
-        public byte Decimals;
+        public UInt160 AssetId { get; }
+        public string AssetName { get; }
+        public string Symbol { get; }
+        public byte Decimals { get; }
 
         public AssetDescriptor(UInt160 asset_id)
         {
+            using SnapshotView snapshot = Blockchain.Singleton.GetSnapshot();
+            var contract = NativeContract.ContractManagement.GetContract(snapshot, asset_id);
+            if (contract is null) throw new ArgumentException();
+
             byte[] script;
             using (ScriptBuilder sb = new ScriptBuilder())
             {
-                sb.EmitAppCall(asset_id, "decimals");
-                sb.EmitAppCall(asset_id, "name");
+                sb.EmitDynamicCall(asset_id, "decimals", true);
+                sb.EmitDynamicCall(asset_id, "symbol", true);
                 script = sb.ToArray();
             }
-            using ApplicationEngine engine = ApplicationEngine.Run(script, gas: 3_000_000);
+            using ApplicationEngine engine = ApplicationEngine.Run(script, snapshot, gas: 0_10000000);
             if (engine.State.HasFlag(VMState.FAULT)) throw new ArgumentException();
             this.AssetId = asset_id;
-            this.AssetName = engine.ResultStack.Pop().GetString();
+            this.AssetName = contract.Manifest.Name;
+            this.Symbol = engine.ResultStack.Pop().GetString();
             this.Decimals = (byte)engine.ResultStack.Pop().GetInteger();
         }
 
