@@ -1,5 +1,6 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Neo.Cryptography.ECC;
+using Neo.IO.Json;
 using Neo.SmartContract.Manifest;
 using System.IO;
 
@@ -11,22 +12,22 @@ namespace Neo.UnitTests.SmartContract.Manifest
         [TestMethod]
         public void ParseFromJson_Default()
         {
-            var json = @"{""groups"":[],""supportedstandards"":[],""abi"":{""hash"":""0x0000000000000000000000000000000000000000"",""methods"":[],""events"":[]},""permissions"":[{""contract"":""*"",""methods"":""*""}],""trusts"":[],""safemethods"":[],""extra"":null}";
+            var json = @"{""name"":""testManifest"",""groups"":[],""supportedstandards"":[],""abi"":{""methods"":[],""events"":[]},""permissions"":[{""contract"":""*"",""methods"":""*""}],""trusts"":[],""extra"":null}";
             var manifest = ContractManifest.Parse(json);
 
             Assert.AreEqual(manifest.ToString(), json);
-            Assert.AreEqual(manifest.ToString(), TestUtils.CreateDefaultManifest(UInt160.Zero).ToString());
+            Assert.AreEqual(manifest.ToString(), TestUtils.CreateDefaultManifest().ToString());
             Assert.IsTrue(manifest.IsValid(UInt160.Zero));
         }
 
         [TestMethod]
         public void ParseFromJson_Permissions()
         {
-            var json = @"{""groups"":[],""supportedstandards"":[],""abi"":{""hash"":""0x0000000000000000000000000000000000000000"",""methods"":[],""events"":[]},""permissions"":[{""contract"":""0x0000000000000000000000000000000000000000"",""methods"":[""method1"",""method2""]}],""trusts"":[],""safemethods"":[],""extra"":null}";
+            var json = @"{""name"":""testManifest"",""groups"":[],""supportedstandards"":[],""abi"":{""methods"":[],""events"":[]},""permissions"":[{""contract"":""0x0000000000000000000000000000000000000000"",""methods"":[""method1"",""method2""]}],""trusts"":[],""extra"":null}";
             var manifest = ContractManifest.Parse(json);
             Assert.AreEqual(manifest.ToString(), json);
 
-            var check = TestUtils.CreateDefaultManifest(UInt160.Zero);
+            var check = TestUtils.CreateDefaultManifest();
             check.Permissions = new[]
             {
                 new ContractPermission()
@@ -41,23 +42,22 @@ namespace Neo.UnitTests.SmartContract.Manifest
         [TestMethod]
         public void ParseFromJson_SafeMethods()
         {
-            var json = @"{""groups"":[],""supportedstandards"":[],""abi"":{""hash"":""0x0000000000000000000000000000000000000000"",""methods"":[],""events"":[]},""permissions"":[{""contract"":""*"",""methods"":""*""}],""trusts"":[],""safemethods"":[""balanceOf""],""extra"":null}";
+            var json = @"{""name"":""testManifest"",""groups"":[],""supportedstandards"":[],""abi"":{""methods"":[],""events"":[]},""permissions"":[{""contract"":""*"",""methods"":""*""}],""trusts"":[],""extra"":null}";
             var manifest = ContractManifest.Parse(json);
             Assert.AreEqual(manifest.ToString(), json);
 
-            var check = TestUtils.CreateDefaultManifest(UInt160.Zero);
-            check.SafeMethods = WildcardContainer<string>.Create("balanceOf");
+            var check = TestUtils.CreateDefaultManifest();
             Assert.AreEqual(manifest.ToString(), check.ToString());
         }
 
         [TestMethod]
         public void ParseFromJson_Trust()
         {
-            var json = @"{""groups"":[],""supportedstandards"":[],""abi"":{""hash"":""0x0000000000000000000000000000000000000000"",""methods"":[],""events"":[]},""permissions"":[{""contract"":""*"",""methods"":""*""}],""trusts"":[""0x0000000000000000000000000000000000000001""],""safemethods"":[],""extra"":null}";
+            var json = @"{""name"":""testManifest"",""groups"":[],""supportedstandards"":[],""abi"":{""methods"":[],""events"":[]},""permissions"":[{""contract"":""*"",""methods"":""*""}],""trusts"":[""0x0000000000000000000000000000000000000001""],""extra"":null}";
             var manifest = ContractManifest.Parse(json);
             Assert.AreEqual(manifest.ToString(), json);
 
-            var check = TestUtils.CreateDefaultManifest(UInt160.Zero);
+            var check = TestUtils.CreateDefaultManifest();
             check.Trusts = WildcardContainer<UInt160>.Create(UInt160.Parse("0x0000000000000000000000000000000000000001"));
             Assert.AreEqual(manifest.ToString(), check.ToString());
         }
@@ -65,11 +65,11 @@ namespace Neo.UnitTests.SmartContract.Manifest
         [TestMethod]
         public void ParseFromJson_Groups()
         {
-            var json = @"{""groups"":[{""pubkey"":""03b209fd4f53a7170ea4444e0cb0a6bb6a53c2bd016926989cf85f9b0fba17a70c"",""signature"":""QUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQQ==""}],""supportedstandards"":[],""abi"":{""hash"":""0x0000000000000000000000000000000000000000"",""methods"":[],""events"":[]},""permissions"":[{""contract"":""*"",""methods"":""*""}],""trusts"":[],""safemethods"":[],""extra"":null}";
+            var json = @"{""name"":""testManifest"",""groups"":[{""pubkey"":""03b209fd4f53a7170ea4444e0cb0a6bb6a53c2bd016926989cf85f9b0fba17a70c"",""signature"":""QUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQUFBQQ==""}],""supportedstandards"":[],""abi"":{""methods"":[],""events"":[]},""permissions"":[{""contract"":""*"",""methods"":""*""}],""trusts"":[],""extra"":null}";
             var manifest = ContractManifest.Parse(json);
             Assert.AreEqual(manifest.ToString(), json);
 
-            var check = TestUtils.CreateDefaultManifest(UInt160.Zero);
+            var check = TestUtils.CreateDefaultManifest();
             check.Groups = new ContractGroup[] { new ContractGroup() { PubKey = ECPoint.Parse("03b209fd4f53a7170ea4444e0cb0a6bb6a53c2bd016926989cf85f9b0fba17a70c", ECCurve.Secp256r1), Signature = "41414141414141414141414141414141414141414141414141414141414141414141414141414141414141414141414141414141414141414141414141414141".HexToBytes() } };
             Assert.AreEqual(manifest.ToString(), check.ToString());
         }
@@ -77,7 +77,7 @@ namespace Neo.UnitTests.SmartContract.Manifest
         [TestMethod]
         public void ParseFromJson_Extra()
         {
-            var json = @"{""groups"":[],""supportedstandards"":[],""abi"":{""hash"":""0x0000000000000000000000000000000000000000"",""methods"":[],""events"":[]},""permissions"":[{""contract"":""*"",""methods"":""*""}],""trusts"":[],""safemethods"":[],""extra"":{""key"":""value""}}";
+            var json = @"{""name"":""testManifest"",""groups"":[],""supportedstandards"":[],""abi"":{""hash"":""0x0000000000000000000000000000000000000000"",""methods"":[],""events"":[]},""permissions"":[{""contract"":""*"",""methods"":""*""}],""trusts"":[],""extra"":{""key"":""value""}}";
             var manifest = ContractManifest.Parse(json);
             Assert.AreEqual(json, json);
             Assert.AreEqual("value", manifest.Extra["key"].AsString(), false);
@@ -89,28 +89,21 @@ namespace Neo.UnitTests.SmartContract.Manifest
             MemoryStream stream = new MemoryStream();
             BinaryWriter writer = new BinaryWriter(stream);
             BinaryReader reader = new BinaryReader(stream);
-            var expected = TestUtils.CreateDefaultManifest(UInt160.Zero);
-            expected.SafeMethods = WildcardContainer<string>.Create(new string[] { "AAA" });
+            var expected = TestUtils.CreateDefaultManifest();
+            expected.Extra = JObject.Parse(@"{""a"":123}");
             expected.Serialize(writer);
             stream.Seek(0, SeekOrigin.Begin);
-            var actual = TestUtils.CreateDefaultManifest(UInt160.Zero);
+            var actual = TestUtils.CreateDefaultManifest();
             actual.Deserialize(reader);
-            Assert.AreEqual(expected.SafeMethods.ToString(), actual.SafeMethods.ToString());
-            Assert.AreEqual(expected.SafeMethods.Count, 1);
-        }
-
-        [TestMethod]
-        public void TestGetHash()
-        {
-            var temp = TestUtils.CreateDefaultManifest(UInt160.Zero);
-            Assert.AreEqual(temp.Abi.Hash, temp.Hash);
+            Assert.AreEqual(expected.ToString(), actual.ToString());
+            Assert.AreEqual(expected.Extra.ToString(), @"{""a"":123}");
         }
 
         [TestMethod]
         public void TestGetSize()
         {
-            var temp = TestUtils.CreateDefaultManifest(UInt160.Zero);
-            Assert.AreEqual(212, temp.Size);
+            var temp = TestUtils.CreateDefaultManifest();
+            Assert.AreEqual(165, temp.Size);
         }
 
         [TestMethod]
@@ -121,18 +114,10 @@ namespace Neo.UnitTests.SmartContract.Manifest
         }
 
         [TestMethod]
-        public void TestCanCall()
-        {
-            var temp = TestUtils.CreateDefaultManifest(UInt160.Zero);
-            temp.SafeMethods = WildcardContainer<string>.Create(new string[] { "AAA" });
-            Assert.AreEqual(true, temp.CanCall(TestUtils.CreateDefaultManifest(UInt160.Zero), "AAA"));
-        }
-
-        [TestMethod]
         public void TestClone()
         {
-            var expected = TestUtils.CreateDefaultManifest(UInt160.Zero);
-            expected.SafeMethods = WildcardContainer<string>.Create(new string[] { "AAA" });
+            var expected = TestUtils.CreateDefaultManifest();
+            expected.Extra = JObject.Parse(@"{ ""a"":123}");
             var actual = expected.Clone();
             Assert.AreEqual(actual.ToString(), expected.ToString());
         }
