@@ -2,6 +2,7 @@ using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Neo.Ledger;
 using Neo.SmartContract;
+using Neo.UnitTests.Extensions;
 using Neo.VM;
 
 namespace Neo.UnitTests.SmartContract
@@ -23,7 +24,7 @@ namespace Neo.UnitTests.SmartContract
             using (ApplicationEngine ae = ApplicationEngine.Create(TriggerType.Application, null, null, 0))
             {
                 ae.LoadScript(SyscallSystemRuntimeCheckWitnessHash);
-                ApplicationEngine.System_Runtime_CheckWitness.FixedPrice.Should().Be(0_00030000L);
+                ApplicationEngine.System_Runtime_CheckWitness.FixedPrice.Should().Be(0_00001024L);
             }
 
             // System.Storage.GetContext: 9bf667ce (price is 1)
@@ -31,7 +32,7 @@ namespace Neo.UnitTests.SmartContract
             using (ApplicationEngine ae = ApplicationEngine.Create(TriggerType.Application, null, null, 0))
             {
                 ae.LoadScript(SyscallSystemStorageGetContextHash);
-                ApplicationEngine.System_Storage_GetContext.FixedPrice.Should().Be(0_00000400L);
+                ApplicationEngine.System_Storage_GetContext.FixedPrice.Should().Be(0_00000016L);
             }
 
             // System.Storage.Get: 925de831 (price is 100)
@@ -39,7 +40,7 @@ namespace Neo.UnitTests.SmartContract
             using (ApplicationEngine ae = ApplicationEngine.Create(TriggerType.Application, null, null, 0))
             {
                 ae.LoadScript(SyscallSystemStorageGetHash);
-                ApplicationEngine.System_Storage_Get.FixedPrice.Should().Be(0_01000000L);
+                ApplicationEngine.System_Storage_Get.FixedPrice.Should().Be(32768L);
             }
         }
 
@@ -61,7 +62,7 @@ namespace Neo.UnitTests.SmartContract
 
             var snapshot = Blockchain.Singleton.GetSnapshot();
             snapshot.Storages.Add(skey, sItem);
-            snapshot.Contracts.Add(script.ToScriptHash(), contractState);
+            snapshot.AddContract(script.ToScriptHash(), contractState);
 
             using (ApplicationEngine ae = ApplicationEngine.Create(TriggerType.Application, null, snapshot))
             {
@@ -72,7 +73,7 @@ namespace Neo.UnitTests.SmartContract
                 debugger.StepInto();
                 var setupPrice = ae.GasConsumed;
                 debugger.Execute();
-                (ae.GasConsumed - setupPrice).Should().Be(ApplicationEngine.StoragePrice * (1 + value.Length));
+                (ae.GasConsumed - setupPrice).Should().Be(ae.StoragePrice * (1 + value.Length));
             }
         }
 
@@ -94,7 +95,7 @@ namespace Neo.UnitTests.SmartContract
 
             var snapshot = Blockchain.Singleton.GetSnapshot();
             snapshot.Storages.Add(skey, sItem);
-            snapshot.Contracts.Add(script.ToScriptHash(), contractState);
+            snapshot.AddContract(script.ToScriptHash(), contractState);
 
             using (ApplicationEngine applicationEngine = ApplicationEngine.Create(TriggerType.Application, null, snapshot))
             {
@@ -105,7 +106,7 @@ namespace Neo.UnitTests.SmartContract
                 debugger.StepInto();
                 var setupPrice = applicationEngine.GasConsumed;
                 debugger.Execute();
-                (applicationEngine.GasConsumed - setupPrice).Should().Be(1 * ApplicationEngine.StoragePrice);
+                (applicationEngine.GasConsumed - setupPrice).Should().Be(1 * applicationEngine.StoragePrice);
             }
         }
 
@@ -129,7 +130,7 @@ namespace Neo.UnitTests.SmartContract
 
             var snapshot = Blockchain.Singleton.GetSnapshot();
             snapshot.Storages.Add(skey, sItem);
-            snapshot.Contracts.Add(script.ToScriptHash(), contractState);
+            snapshot.AddContract(script.ToScriptHash(), contractState);
 
             using (ApplicationEngine ae = ApplicationEngine.Create(TriggerType.Application, null, snapshot))
             {
@@ -141,7 +142,7 @@ namespace Neo.UnitTests.SmartContract
                 var setupPrice = ae.GasConsumed;
                 debugger.StepInto();
                 debugger.StepInto();
-                (ae.GasConsumed - setupPrice).Should().Be((1 + (oldValue.Length / 4) + value.Length - oldValue.Length) * ApplicationEngine.StoragePrice);
+                (ae.GasConsumed - setupPrice).Should().Be((1 + (oldValue.Length / 4) + value.Length - oldValue.Length) * ae.StoragePrice);
             }
         }
 
@@ -165,7 +166,7 @@ namespace Neo.UnitTests.SmartContract
 
             var snapshot = Blockchain.Singleton.GetSnapshot();
             snapshot.Storages.Add(skey, sItem);
-            snapshot.Contracts.Add(script.ToScriptHash(), contractState);
+            snapshot.AddContract(script.ToScriptHash(), contractState);
 
             using (ApplicationEngine ae = ApplicationEngine.Create(TriggerType.Application, null, snapshot))
             {
@@ -180,7 +181,7 @@ namespace Neo.UnitTests.SmartContract
                 debugger.StepInto(); //syscall Storage.GetContext
                 var setupPrice = ae.GasConsumed;
                 debugger.StepInto(); //syscall Storage.Put
-                (ae.GasConsumed - setupPrice).Should().Be((sItem.Value.Length / 4 + 1) * ApplicationEngine.StoragePrice); // = PUT basic fee
+                (ae.GasConsumed - setupPrice).Should().Be((sItem.Value.Length / 4 + 1) * ae.StoragePrice); // = PUT basic fee
             }
         }
 
