@@ -8,6 +8,7 @@ using Neo.Network.P2P;
 using Neo.Network.P2P.Payloads;
 using Neo.SmartContract;
 using Neo.SmartContract.Manifest;
+using Neo.UnitTests.Extensions;
 using Neo.VM;
 using Neo.VM.Types;
 using Neo.Wallets;
@@ -48,11 +49,12 @@ namespace Neo.UnitTests.SmartContract
 
                 scriptHash2 = script.ToArray().ToScriptHash();
 
-                snapshot.Contracts.Delete(scriptHash2);
-                snapshot.Contracts.Add(scriptHash2, new ContractState()
+                snapshot.DeleteContract(scriptHash2);
+                snapshot.AddContract(scriptHash2, new ContractState()
                 {
-                    Script = script.ToArray(),
-                    Manifest = TestUtils.CreateManifest(scriptHash2, "test", ContractParameterType.Any, ContractParameterType.Integer, ContractParameterType.Integer),
+                    Nef = new NefFile { Script = script.ToArray() },
+                    Hash = script.ToArray().ToScriptHash(),
+                    Manifest = TestUtils.CreateManifest("test", ContractParameterType.Any, ContractParameterType.Integer, ContractParameterType.Integer),
                 });
             }
 
@@ -87,7 +89,7 @@ namespace Neo.UnitTests.SmartContract
 
                 // Call script
 
-                script.EmitAppCall(scriptHash2, "test", "testEvent2", 1);
+                script.EmitDynamicCall(scriptHash2, "test", true, "testEvent2", 1);
 
                 // Drop return
 
@@ -139,7 +141,7 @@ namespace Neo.UnitTests.SmartContract
 
                 // Call script
 
-                script.EmitAppCall(scriptHash2, "test", "testEvent2", 1);
+                script.EmitDynamicCall(scriptHash2, "test", true, "testEvent2", 1);
 
                 // Drop return
 
@@ -178,7 +180,7 @@ namespace Neo.UnitTests.SmartContract
 
             // Clean storage
 
-            snapshot.Contracts.Delete(scriptHash2);
+            snapshot.DeleteContract(scriptHash2);
         }
 
         private void AssertNotification(StackItem stackItem, UInt160 scriptHash, string notification)
@@ -214,14 +216,15 @@ namespace Neo.UnitTests.SmartContract
 
             var contract = new ContractState()
             {
-                Manifest = TestUtils.CreateManifest(scriptA.ToArray().ToScriptHash(), "test", ContractParameterType.Any, ContractParameterType.Integer, ContractParameterType.Integer),
-                Script = scriptA.ToArray()
+                Manifest = TestUtils.CreateManifest("test", ContractParameterType.Any, ContractParameterType.Integer, ContractParameterType.Integer),
+                Nef = new NefFile { Script = scriptA.ToArray() },
+                Hash = scriptA.ToArray().ToScriptHash()
             };
             engine = GetEngine(true, true, false);
-            engine.Snapshot.Contracts.Add(contract.ScriptHash, contract);
+            engine.Snapshot.AddContract(contract.Hash, contract);
 
             using ScriptBuilder scriptB = new ScriptBuilder();
-            scriptB.EmitAppCall(contract.ScriptHash, "test", 0, 1);
+            scriptB.EmitDynamicCall(contract.Hash, "test", true, 0, 1);
             engine.LoadScript(scriptB.ToArray());
 
             Assert.AreEqual(VMState.HALT, engine.Execute());
@@ -385,14 +388,14 @@ namespace Neo.UnitTests.SmartContract
                                         0x01, 0x01, 0x01, 0x01, 0x01,
                                         0x01, 0x01, 0x01, 0x01, 0x01,
                                         0x01, 0x01, 0x01, 0x01, 0x01 };
-            engine.GetContract(new UInt160(data1)).Should().BeNull();
+            Neo.SmartContract.Native.NativeContract.ContractManagement.GetContract(engine.Snapshot, new UInt160(data1)).Should().BeNull();
 
             var snapshot = Blockchain.Singleton.GetSnapshot();
             var state = TestUtils.GetContract();
-            snapshot.Contracts.Add(state.ScriptHash, state);
+            snapshot.AddContract(state.Hash, state);
             engine = ApplicationEngine.Create(TriggerType.Application, null, snapshot);
             engine.LoadScript(new byte[] { 0x01 });
-            engine.GetContract(state.ScriptHash).Should().BeSameAs(state);
+            Neo.SmartContract.Native.NativeContract.ContractManagement.GetContract(engine.Snapshot, state.Hash).Should().BeSameAs(state);
         }
 
         [TestMethod]
@@ -400,7 +403,7 @@ namespace Neo.UnitTests.SmartContract
         {
             var engine = GetEngine(false, true);
             var state = TestUtils.GetContract();
-            engine.Snapshot.Contracts.Add(state.ScriptHash, state);
+            engine.Snapshot.AddContract(state.Hash, state);
             engine.LoadScript(state.Script);
             engine.GetStorageContext().IsReadOnly.Should().BeFalse();
         }
@@ -410,7 +413,7 @@ namespace Neo.UnitTests.SmartContract
         {
             var engine = GetEngine(false, true);
             var state = TestUtils.GetContract();
-            engine.Snapshot.Contracts.Add(state.ScriptHash, state);
+            engine.Snapshot.AddContract(state.Hash, state);
             engine.LoadScript(state.Script);
             engine.GetReadOnlyContext().IsReadOnly.Should().BeTrue();
         }
@@ -432,7 +435,7 @@ namespace Neo.UnitTests.SmartContract
                 Value = new byte[] { 0x01, 0x02, 0x03, 0x04 },
                 IsConstant = true
             };
-            snapshot.Contracts.Add(state.ScriptHash, state);
+            snapshot.AddContract(state.Hash, state);
             snapshot.Storages.Add(storageKey, storageItem);
             var engine = ApplicationEngine.Create(TriggerType.Application, null, snapshot);
             engine.LoadScript(new byte[] { 0x01 });
@@ -489,7 +492,7 @@ namespace Neo.UnitTests.SmartContract
                 Value = new byte[] { 0x01, 0x02, 0x03, 0x04 },
                 IsConstant = true
             };
-            snapshot.Contracts.Add(state.ScriptHash, state);
+            snapshot.AddContract(state.Hash, state);
             snapshot.Storages.Add(storageKey, storageItem);
             engine = ApplicationEngine.Create(TriggerType.Application, null, snapshot);
             engine.LoadScript(new byte[] { 0x01 });
@@ -523,7 +526,7 @@ namespace Neo.UnitTests.SmartContract
                 Value = new byte[] { 0x01, 0x02, 0x03, 0x04 },
                 IsConstant = false
             };
-            snapshot.Contracts.Add(state.ScriptHash, state);
+            snapshot.AddContract(state.Hash, state);
             snapshot.Storages.Add(storageKey, storageItem);
             var engine = ApplicationEngine.Create(TriggerType.Application, null, snapshot);
             engine.LoadScript(new byte[] { 0x01 });
@@ -553,7 +556,7 @@ namespace Neo.UnitTests.SmartContract
                 Value = new byte[] { 0x01, 0x02, 0x03, 0x04 },
                 IsConstant = false
             };
-            snapshot.Contracts.Add(state.ScriptHash, state);
+            snapshot.AddContract(state.Hash, state);
             snapshot.Storages.Add(storageKey, storageItem);
             engine = ApplicationEngine.Create(TriggerType.Application, null, snapshot);
             engine.LoadScript(new byte[] { 0x01 });
@@ -591,60 +594,33 @@ namespace Neo.UnitTests.SmartContract
             var args = new VM.Types.Array { 0, 1 };
             var state = TestUtils.GetContract(method, args.Count);
 
-            snapshot.Contracts.Add(state.ScriptHash, state);
+            snapshot.AddContract(state.Hash, state);
             var engine = ApplicationEngine.Create(TriggerType.Application, null, snapshot);
             engine.LoadScript(new byte[] { 0x01 });
 
-            engine.CallContract(state.ScriptHash, method, args);
+            engine.Push(args[1]); engine.Push(args[0]);
+            engine.CallContract(state.Hash, method, CallFlags.All, false, (ushort)args.Count);
             engine.CurrentContext.EvaluationStack.Pop().Should().Be(args[0]);
             engine.CurrentContext.EvaluationStack.Pop().Should().Be(args[1]);
 
             state.Manifest.Permissions[0].Methods = WildcardContainer<string>.Create("a");
-            Assert.ThrowsException<InvalidOperationException>(() => engine.CallContract(state.ScriptHash, method, args));
+            engine.Push(args[1]); engine.Push(args[0]);
+            Assert.ThrowsException<InvalidOperationException>(() => engine.CallContract(state.Hash, method, CallFlags.All, false, (ushort)args.Count));
 
             state.Manifest.Permissions[0].Methods = WildcardContainer<string>.CreateWildcard();
-            engine.CallContract(state.ScriptHash, method, args);
+            engine.Push(args[1]); engine.Push(args[0]);
+            engine.CallContract(state.Hash, method, CallFlags.All, false, (ushort)args.Count);
 
-            Assert.ThrowsException<InvalidOperationException>(() => engine.CallContract(UInt160.Zero, method, args));
-        }
-
-        [TestMethod]
-        public void TestContract_CallEx()
-        {
-            var snapshot = Blockchain.Singleton.GetSnapshot();
-
-            string method = "method";
-            var args = new VM.Types.Array { 0, 1 };
-            var state = TestUtils.GetContract(method, args.Count);
-            snapshot.Contracts.Add(state.ScriptHash, state);
-
-
-            foreach (var flags in new CallFlags[] { CallFlags.None, CallFlags.AllowCall, CallFlags.AllowModifyStates, CallFlags.All })
-            {
-                var engine = ApplicationEngine.Create(TriggerType.Application, null, snapshot);
-                engine.LoadScript(new byte[] { 0x01 });
-
-                engine.CallContractEx(state.ScriptHash, method, args, CallFlags.All);
-                engine.CurrentContext.EvaluationStack.Pop().Should().Be(args[0]);
-                engine.CurrentContext.EvaluationStack.Pop().Should().Be(args[1]);
-
-                // Contract doesn't exists
-                Assert.ThrowsException<InvalidOperationException>(() => engine.CallContractEx(UInt160.Zero, method, args, CallFlags.All));
-
-                // Call with rights
-                engine.CallContractEx(state.ScriptHash, method, args, flags);
-                engine.CurrentContext.EvaluationStack.Pop().Should().Be(args[0]);
-                engine.CurrentContext.EvaluationStack.Pop().Should().Be(args[1]);
-            }
+            engine.Push(args[1]); engine.Push(args[0]);
+            Assert.ThrowsException<InvalidOperationException>(() => engine.CallContract(UInt160.Zero, method, CallFlags.All, false, (ushort)args.Count));
         }
 
         [TestMethod]
         public void TestContract_Destroy()
         {
-            var engine = GetEngine(false, true);
-            engine.DestroyContract();
+            var snapshot = Blockchain.Singleton.GetSnapshot().Clone();
+            snapshot.PersistingBlock = new Block();
 
-            var snapshot = Blockchain.Singleton.GetSnapshot();
             var state = TestUtils.GetContract();
             var scriptHash = UInt160.Parse("0xcb9f3b7c6fb1cf2c13a40637c189bdd066a272b4");
             var storageItem = new StorageItem
@@ -658,21 +634,16 @@ namespace Neo.UnitTests.SmartContract
                 Id = 0x43000000,
                 Key = new byte[] { 0x01 }
             };
-            snapshot.Contracts.Add(scriptHash, state);
+            snapshot.AddContract(scriptHash, state);
             snapshot.Storages.Add(storageKey, storageItem);
-            engine = ApplicationEngine.Create(TriggerType.Application, null, snapshot);
-            engine.LoadScript(new byte[0]);
-            engine.DestroyContract();
-            engine.Snapshot.Storages.Find(BitConverter.GetBytes(0x43000000)).Any().Should().BeFalse();
+            snapshot.DestroyContract(scriptHash);
+            snapshot.Storages.Find(BitConverter.GetBytes(0x43000000)).Any().Should().BeFalse();
 
             //storages are removed
-            snapshot = Blockchain.Singleton.GetSnapshot();
             state = TestUtils.GetContract();
-            snapshot.Contracts.Add(scriptHash, state);
-            engine = ApplicationEngine.Create(TriggerType.Application, null, snapshot);
-            engine.LoadScript(new byte[0]);
-            engine.DestroyContract();
-            engine.Snapshot.Storages.Find(BitConverter.GetBytes(0x43000000)).Any().Should().BeFalse();
+            snapshot.AddContract(scriptHash, state);
+            snapshot.DestroyContract(scriptHash);
+            snapshot.Storages.Find(BitConverter.GetBytes(0x43000000)).Any().Should().BeFalse();
         }
 
         [TestMethod]
@@ -689,26 +660,26 @@ namespace Neo.UnitTests.SmartContract
             tx.Script = new byte[] { 0x01, 0x02, 0x03 };
         }
 
-        private static ApplicationEngine GetEngine(bool hasContainer = false, bool hasSnapshot = false, bool addScript = true)
+        private static ApplicationEngine GetEngine(bool hasContainer = false, bool hasSnapshot = false, bool addScript = true, long gas = 20_00000000)
         {
             var tx = TestUtils.GetTransaction(UInt160.Zero);
-            var snapshot = Blockchain.Singleton.GetSnapshot();
+            var snapshot = Blockchain.Singleton.GetSnapshot().Clone();
             ApplicationEngine engine;
             if (hasContainer && hasSnapshot)
             {
-                engine = ApplicationEngine.Create(TriggerType.Application, tx, snapshot);
+                engine = ApplicationEngine.Create(TriggerType.Application, tx, snapshot, gas);
             }
             else if (hasContainer && !hasSnapshot)
             {
-                engine = ApplicationEngine.Create(TriggerType.Application, tx, null);
+                engine = ApplicationEngine.Create(TriggerType.Application, tx, null, gas);
             }
             else if (!hasContainer && hasSnapshot)
             {
-                engine = ApplicationEngine.Create(TriggerType.Application, null, snapshot);
+                engine = ApplicationEngine.Create(TriggerType.Application, null, snapshot, gas);
             }
             else
             {
-                engine = ApplicationEngine.Create(TriggerType.Application, null, null);
+                engine = ApplicationEngine.Create(TriggerType.Application, null, null, gas);
             }
             if (addScript)
             {
