@@ -50,74 +50,116 @@ namespace Neo.UnitTests.VMT
         [TestMethod]
         public void TestEmitAppCall1()
         {
-            //format:(byte)0x10+(byte)OpCode.NEWARRAY+(string)operation+(Uint160)scriptHash+(uint)InteropService.System_Contract_Call
             ScriptBuilder sb = new ScriptBuilder();
-            sb.EmitAppCall(UInt160.Zero, "AAAAA");
-            byte[] tempArray = new byte[36];
+            sb.EmitDynamicCall(UInt160.Zero, "AAAAA", false);
+            byte[] tempArray = new byte[37];
             tempArray[0] = (byte)OpCode.PUSH0;
-            tempArray[1] = (byte)OpCode.NEWARRAY;
-            tempArray[2] = (byte)OpCode.PUSHDATA1;
-            tempArray[3] = 5;//operation.Length
-            Array.Copy(Encoding.UTF8.GetBytes("AAAAA"), 0, tempArray, 4, 5);//operation.data
-            tempArray[9] = (byte)OpCode.PUSHDATA1;
-            tempArray[10] = 0x14;//scriptHash.Length
-            Array.Copy(UInt160.Zero.ToArray(), 0, tempArray, 11, 20);//operation.data
+            tempArray[1] = (byte)OpCode.PUSH0;
+            tempArray[2] = (byte)OpCode.PUSH15;//(byte)CallFlags.All;
+            tempArray[3] = (byte)OpCode.PUSHDATA1;
+            tempArray[4] = 5;//operation.Length
+            Array.Copy(Encoding.UTF8.GetBytes("AAAAA"), 0, tempArray, 5, 5);//operation.data
+            tempArray[10] = (byte)OpCode.PUSHDATA1;
+            tempArray[11] = 0x14;//scriptHash.Length
+            Array.Copy(UInt160.Zero.ToArray(), 0, tempArray, 12, 20);//operation.data
             uint api = ApplicationEngine.System_Contract_Call;
-            tempArray[31] = (byte)OpCode.SYSCALL;
-            Array.Copy(BitConverter.GetBytes(api), 0, tempArray, 32, 4);//api.data
-            CollectionAssert.AreEqual(tempArray, sb.ToArray());
+            tempArray[32] = (byte)OpCode.SYSCALL;
+            Array.Copy(BitConverter.GetBytes(api), 0, tempArray, 33, 4);//api.data
+            Assert.AreEqual(tempArray.ToHexString(), sb.ToArray().ToHexString());
+        }
+
+        [TestMethod]
+        public void TestEmitArray()
+        {
+            var expected = new BigInteger[] { 1, 2, 3 };
+            var sb = new ScriptBuilder();
+            sb.CreateArray(expected);
+
+            using var engine = ApplicationEngine.Create(TriggerType.Application, null, null);
+            engine.LoadScript(sb.ToArray());
+            Assert.AreEqual(VMState.HALT, engine.Execute());
+
+            CollectionAssert.AreEqual(expected, engine.ResultStack.Pop<VM.Types.Array>().Select(u => u.GetInteger()).ToArray());
+
+            expected = new BigInteger[] { };
+            sb = new ScriptBuilder();
+            sb.CreateArray(expected);
+
+            using var engine2 = ApplicationEngine.Create(TriggerType.Application, null, null);
+            engine2.LoadScript(sb.ToArray());
+            Assert.AreEqual(VMState.HALT, engine2.Execute());
+
+            Assert.AreEqual(0, engine2.ResultStack.Pop<VM.Types.Array>().Count);
+        }
+
+        [TestMethod]
+        public void TestEmitMap()
+        {
+            var expected = new Dictionary<BigInteger, BigInteger>() { { 1, 2 }, { 3, 4 } };
+            var sb = new ScriptBuilder();
+            sb.CreateMap(expected);
+
+            using var engine = ApplicationEngine.Create(TriggerType.Application, null, null);
+            engine.LoadScript(sb.ToArray());
+            Assert.AreEqual(VMState.HALT, engine.Execute());
+
+            var map = engine.ResultStack.Pop<VM.Types.Map>();
+            var dic = map.ToDictionary(u => u.Key, u => u.Value);
+
+            CollectionAssert.AreEqual(expected.Keys, dic.Keys.Select(u => u.GetInteger()).ToArray());
+            CollectionAssert.AreEqual(expected.Values, dic.Values.Select(u => u.GetInteger()).ToArray());
         }
 
         [TestMethod]
         public void TestEmitAppCall2()
         {
-            //format:(ContractParameter[])ContractParameter+(byte)OpCode.PACK+(string)operation+(Uint160)scriptHash+(uint)InteropService.System_Contract_Call
             ScriptBuilder sb = new ScriptBuilder();
-            sb.EmitAppCall(UInt160.Zero, "AAAAA", new ContractParameter[] { new ContractParameter(ContractParameterType.Integer) });
-            byte[] tempArray = new byte[37];
+            sb.EmitDynamicCall(UInt160.Zero, "AAAAA", false, new ContractParameter[] { new ContractParameter(ContractParameterType.Integer) });
+            byte[] tempArray = new byte[38];
             tempArray[0] = (byte)OpCode.PUSH0;
             tempArray[1] = (byte)OpCode.PUSH1;
-            tempArray[2] = (byte)OpCode.PACK;
-            tempArray[3] = (byte)OpCode.PUSHDATA1;
-            tempArray[4] = 0x05;//operation.Length
-            Array.Copy(Encoding.UTF8.GetBytes("AAAAA"), 0, tempArray, 5, 5);//operation.data
-            tempArray[10] = (byte)OpCode.PUSHDATA1;
-            tempArray[11] = 0x14;//scriptHash.Length
-            Array.Copy(UInt160.Zero.ToArray(), 0, tempArray, 12, 20);//operation.data
+            tempArray[2] = (byte)OpCode.PUSH0;
+            tempArray[3] = (byte)OpCode.PUSH15;//(byte)CallFlags.All;
+            tempArray[4] = (byte)OpCode.PUSHDATA1;
+            tempArray[5] = 0x05;//operation.Length
+            Array.Copy(Encoding.UTF8.GetBytes("AAAAA"), 0, tempArray, 6, 5);//operation.data
+            tempArray[11] = (byte)OpCode.PUSHDATA1;
+            tempArray[12] = 0x14;//scriptHash.Length
+            Array.Copy(UInt160.Zero.ToArray(), 0, tempArray, 13, 20);//operation.data
             uint api = ApplicationEngine.System_Contract_Call;
-            tempArray[32] = (byte)OpCode.SYSCALL;
-            Array.Copy(BitConverter.GetBytes(api), 0, tempArray, 33, 4);//api.data
-            CollectionAssert.AreEqual(tempArray, sb.ToArray());
+            tempArray[33] = (byte)OpCode.SYSCALL;
+            Array.Copy(BitConverter.GetBytes(api), 0, tempArray, 34, 4);//api.data
+            Assert.AreEqual(tempArray.ToHexString(), sb.ToArray().ToHexString());
         }
 
         [TestMethod]
         public void TestEmitAppCall3()
         {
-            //format:(object[])args+(byte)OpCode.PACK+(string)operation+(Uint160)scriptHash+(uint)InteropService.System_Contract_Call
             ScriptBuilder sb = new ScriptBuilder();
-            sb.EmitAppCall(UInt160.Zero, "AAAAA", true);
-            byte[] tempArray = new byte[37];
-            tempArray[0] = (byte)OpCode.PUSH1;//arg
-            tempArray[1] = (byte)OpCode.PUSH1;//args.Length 
-            tempArray[2] = (byte)OpCode.PACK;
-            tempArray[3] = (byte)OpCode.PUSHDATA1;
-            tempArray[4] = 0x05;//operation.Length
-            Array.Copy(Encoding.UTF8.GetBytes("AAAAA"), 0, tempArray, 5, 5);//operation.data
-            tempArray[10] = (byte)OpCode.PUSHDATA1;
-            tempArray[11] = 0x14;//scriptHash.Length
-            Array.Copy(UInt160.Zero.ToArray(), 0, tempArray, 12, 20);//operation.data
+            sb.EmitDynamicCall(UInt160.Zero, "AAAAA", false, true);
+            byte[] tempArray = new byte[38];
+            tempArray[0] = (byte)OpCode.PUSH1;
+            tempArray[1] = (byte)OpCode.PUSH1;//arg.Length
+            tempArray[2] = (byte)OpCode.PUSH0;//return 
+            tempArray[3] = (byte)OpCode.PUSH15;//(byte)CallFlags.All;
+            tempArray[4] = (byte)OpCode.PUSHDATA1;
+            tempArray[5] = 0x05;//operation.Length
+            Array.Copy(Encoding.UTF8.GetBytes("AAAAA"), 0, tempArray, 6, 5);//operation.data
+            tempArray[11] = (byte)OpCode.PUSHDATA1;
+            tempArray[12] = 0x14;//scriptHash.Length
+            Array.Copy(UInt160.Zero.ToArray(), 0, tempArray, 13, 20);//operation.data
             uint api = ApplicationEngine.System_Contract_Call;
-            tempArray[32] = (byte)OpCode.SYSCALL;
-            Array.Copy(BitConverter.GetBytes(api), 0, tempArray, 33, 4);//api.data
-            CollectionAssert.AreEqual(tempArray, sb.ToArray());
+            tempArray[33] = (byte)OpCode.SYSCALL;
+            Array.Copy(BitConverter.GetBytes(api), 0, tempArray, 34, 4);//api.data
+            Assert.AreEqual(tempArray.ToHexString(), sb.ToArray().ToHexString());
         }
 
         [TestMethod]
         public void TestMakeScript()
         {
-            byte[] testScript = NativeContract.GAS.Hash.MakeScript("balanceOf", UInt160.Zero);
+            byte[] testScript = NativeContract.GAS.Hash.MakeScript("balanceOf", true, UInt160.Zero);
 
-            Assert.AreEqual("0c14000000000000000000000000000000000000000011c00c0962616c616e63654f660c14bcaf41d684c7d4ad6ee0d99da9707b9d1f0c8e6641627d5b52",
+            Assert.AreEqual("0c14000000000000000000000000000000000000000011111f0c0962616c616e63654f660c141717ddafdd757eec365865b963473beb617f9a1441627d5b52",
                             testScript.ToHexString());
         }
 
